move json schemas to restapi app
